feat: moved scopes to a new file
